% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe-calculations.R
\name{calculateRSSNormalised}
\alias{calculateRSSNormalised}
\title{Calculate residual some of squares for n-phase exponential curve}
\usage{
calculateRSSNormalised(lambda, k, data, time_variable = "t",
  value_variable = "value")
}
\arguments{
\item{lambda}{Vector specifying the initial size of each group.}

\item{k}{Vector specifying the decay rates of each group.}

\item{data}{Data frame containing the observations.}

\item{time_variable}{Quoted string specifying the name of the time variable in \code{data}.}

\item{value_variable}{Quoted string specifying the name of the value variable in \code{data}.}
}
\value{
The residual sum of squared error.
}
\description{
Calculates the residual sum of squares for an n-phase exponential curve with
initial sizes, \eqn{\lambda} and rate parameters, \eqn{k}, for a given set of data. \cr\cr
Note importantly that the rate paramters, \eqn{k}, are \strong{not} negated.
}
