% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npe-fitting.R
\name{fitnphaseexponential}
\alias{fitnphaseexponential}
\title{Fit an n-phase exponential distribution}
\usage{
fitnphaseexponential(num_phases = 1, data, time_variable = "t",
  value_variable = "dna", lambda_0, k_0)
}
\arguments{
\item{num_phases}{Number of phases for exponential distributions.}

\item{data}{Data frame containing observations at some time points.}

\item{time_variable}{Quoted string specifying the name of the time variable in \code{data}.}

\item{value_variable}{Quoted string specifying the name of the value variable in \code{data}.}

\item{lambda_0}{\emph{Optional} Initial guess for the size of \eqn{\lambda}.}

\item{k_0}{\emph{Optional} Initial guess for the size of \eqn{k}.}
}
\value{
A list of class "nphasefit", which contains the following elements:
\itemize{
  \item \code{estimates}: Data frame with two columns: \code{lambda} and \code{k}, containing the optimal estimates of \eqn{\lambda} and \eqn{k} respectively.
  \item \code{error}: The residual sum of squares at the optimum.
  \item \code{AIC}: The AIC at the optimum.
}
}
\description{
Fits an n-phase distribution
}
\details{
Under the hood, this function is optimising the residual sum of squares (see \code{\link{calculateRSS}})
using \code{\link[stats]{nlm}}. The only control parameter passed to the optimiser is \code{typsize},
which is the same as the initial guess.

If not supplied, the initial guess for the initial sizes, \eqn{\lambda}, is 200/\code{num_phases}.

If not supplied, the initial guess for the rate parameters, \eqn{k}, is \eqn{10^{-2:-2-(num_phases+1)}}.
That is, \eqn{k_0 = 0.01}, and each consecutive \eqn{k} decreases by an order of magnitude
from there. These guesses seem to work generally well, but may need tweaking for specific problems.
}
